Split providers into modules
