+++ src/lib.rs
@@ -1,4 +1,2 @@
-pub mod cache;
-pub mod data;
-pub mod helper;
-pub mod provider;
+pub mod app;
+pub mod modules2;
