+++ src/modules2/mod.rs
@@ -1,39 +1,16 @@
+// Provider Module
+pub mod archlinux;
+pub mod currency;
+pub mod nsfw;
+
+// Data Module
+pub mod cache;
+pub mod http;
+
+// Every module should provide a function that turn user input to [`Sendable`]
 use reqwest::IntoUrl;
 use std::fmt::Display;
-use std::ops::Deref;
-use std::sync::Arc;
 use teloxide::types::InputFile;
-
-use crate::cache::Cacher;
-use crate::provider::HttpClient;
-
-pub struct AppData(Arc<RuntimeData>);
-
-impl From<RuntimeData> for AppData {
-    fn from(data: RuntimeData) -> Self {
-        Self(Arc::new(data))
-    }
-}
-
-impl Clone for AppData {
-    fn clone(&self) -> Self {
-        AppData(Arc::clone(&self.0))
-    }
-}
-
-impl Deref for AppData {
-    type Target = Arc<RuntimeData>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-#[derive(typed_builder::TypedBuilder)]
-pub struct RuntimeData {
-    pub cacher: Cacher,
-    pub requester: HttpClient,
-}
 
 pub enum Sendable {
     Text(String),
